Add files via upload
This is my first draft of the PCM1690 DAC breakout board. There are a couple notable issues. 1) I used the incorrect footprint, we could update this with the new PCM1690 footprint if we want to order a PCB with no onboard passives. 2) I did not make any copper pours in this draft, draft 3 includes copper ground pours.