Add files via upload
Draft 3 includes some onboard power-supply decoupling capacitors. Based on figure 41, on page 37 of the PCM1690 datasheet, I placed capacitors C1-C3 in this layout. C4-C6 are electrolytic capacitors, I left these off of this breakout board with the intent to connect C4-C6 to the appropriate pins on a breadboard. 

There are a couple issues to fix in this draft. 1) The filled zone for the copper ground pour is not showing up when I try to display it. I followed on online tutorial for adding copper pours, I'm not sure what went wrong. 2) there are a few DRC warnings titled "Track Has Unconnected End" I'm not sure if this is because the tracks are routed from the pins on the DAC footprint to header pins, or if its something else entirely. 3) There are several DRC errors related to clearance. Some are in part due to the fact that I did not modify the courtyard in the footprint for the PCM1690. I think I had the courtyard layer muted what I was laying out some of the tracks, some of the tracks run through the courtyard of the decoupling capacitors. All of these should be relatively easy to fix if we update the courtyard in the new PCM1690 footprint, and route tracks more carefully.